+++ src/main/java/spring/backend/activity/application/QuickStartActivitySelectService.java
@@ -8,8 +8,8 @@ import spring.backend.activity.domain.entity.Activity;
 import spring.backend.activity.domain.repository.ActivityRepository;
 import spring.backend.quickstart.domain.repository.QuickStartRepository;
 import spring.backend.activity.domain.service.FinishActivityAutoService;
-import spring.backend.activity.dto.request.QuickStartActivitySelectRequest;
-import spring.backend.activity.dto.response.QuickStartActivitySelectResponse;
+import spring.backend.activity.presentation.dto.request.QuickStartActivitySelectRequest;
+import spring.backend.activity.presentation.dto.response.QuickStartActivitySelectResponse;
 import spring.backend.activity.exception.ActivityErrorCode;
 import spring.backend.quickstart.exception.QuickStartErrorCode;
 import spring.backend.member.domain.entity.Member;
