+++ src/main/java/spring/backend/activity/presentation/dto/response/ActivitiesByMemberAndKeywordInMonthResponse.java
@@ -1,7 +1,8 @@
-package spring.backend.activity.dto.response;
+package spring.backend.activity.presentation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import spring.backend.activity.domain.value.Keyword;
+import spring.backend.activity.dto.response.ActivityWithTitleAndSavedTimeResponse;
 
 import java.util.List;
