+++ src/main/java/spring/backend/activity/presentation/dto/response/MonthlyActivityOverviewResponse.java
@@ -1,6 +1,8 @@
-package spring.backend.activity.dto.response;
+package spring.backend.activity.presentation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import spring.backend.activity.dto.response.MonthlyActivityCountByKeywordResponse;
+import spring.backend.activity.dto.response.MonthlySavedTimeAndActivityCountResponse;
 
 import java.time.Month;
 import java.util.List;
