+++ src/main/java/spring/backend/activity/presentation/swagger/QuickStartActivitySelectSwagger.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
-import spring.backend.activity.dto.request.QuickStartActivitySelectRequest;
-import spring.backend.activity.dto.response.QuickStartActivitySelectResponse;
+import spring.backend.activity.presentation.dto.request.QuickStartActivitySelectRequest;
+import spring.backend.activity.presentation.dto.response.QuickStartActivitySelectResponse;
 import spring.backend.activity.exception.ActivityErrorCode;
 import spring.backend.quickstart.exception.QuickStartErrorCode;
 import spring.backend.core.configuration.swagger.ApiErrorCode;
