+++ src/main/java/spring/backend/activity/presentation/swagger/ReadActivityCalendarSwagger.java
@@ -5,8 +5,8 @@ import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.http.ResponseEntity;
-import spring.backend.activity.dto.request.ReadActivityCalendarRequest;
-import spring.backend.activity.dto.response.ActivityCalendarResponse;
+import spring.backend.activity.presentation.dto.request.ReadActivityCalendarRequest;
+import spring.backend.activity.presentation.dto.response.ActivityCalendarResponse;
 import spring.backend.activity.exception.ActivityErrorCode;
 import spring.backend.core.configuration.swagger.ApiErrorCode;
 import spring.backend.core.exception.error.GlobalErrorCode;
