+++ src/main/java/spring/backend/activity/presentation/swagger/ReadMonthlyActivityOverviewSwagger.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
-import spring.backend.activity.dto.request.MonthlyActivityOverviewRequest;
-import spring.backend.activity.dto.response.MonthlyActivityOverviewResponse;
+import spring.backend.activity.presentation.dto.request.MonthlyActivityOverviewRequest;
+import spring.backend.activity.presentation.dto.response.MonthlyActivityOverviewResponse;
 import spring.backend.core.presentation.RestResponse;
 import spring.backend.member.domain.entity.Member;
