+++ src/main/java/spring/backend/activity/presentation/swagger/UserActivitySelectSwagger.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
-import spring.backend.activity.dto.request.UserActivitySelectRequest;
-import spring.backend.activity.dto.response.UserActivitySelectResponse;
+import spring.backend.activity.presentation.dto.request.UserActivitySelectRequest;
+import spring.backend.activity.presentation.dto.response.UserActivitySelectResponse;
 import spring.backend.activity.exception.ActivityErrorCode;
 import spring.backend.core.configuration.swagger.ApiErrorCode;
 import spring.backend.core.exception.error.GlobalErrorCode;
