+++ src/test/java/spring/backend/activity/application/QuickStartActivitySelectServiceTest.java
@@ -14,8 +14,8 @@ import spring.backend.quickstart.domain.repository.QuickStartRepository;
 import spring.backend.activity.domain.service.FinishActivityAutoService;
 import spring.backend.activity.domain.value.Keyword;
 import spring.backend.activity.domain.value.Type;
-import spring.backend.activity.dto.request.QuickStartActivitySelectRequest;
-import spring.backend.activity.dto.response.QuickStartActivitySelectResponse;
+import spring.backend.activity.presentation.dto.request.QuickStartActivitySelectRequest;
+import spring.backend.activity.presentation.dto.response.QuickStartActivitySelectResponse;
 import spring.backend.activity.exception.ActivityErrorCode;
 import spring.backend.quickstart.exception.QuickStartErrorCode;
 import spring.backend.core.exception.DomainException;
