+++ src/test/java/spring/backend/activity/domain/repository/ActivityRepositoryTest.java
@@ -11,7 +11,7 @@ import spring.backend.activity.domain.value.Type;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 @SpringBootTest
 class ActivityRepositoryTest {
